# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface VoterRegistration {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 4 */
  voterId: string;
  email: string;
  mobile: string;
  /** @minLength 8 */
  password: string;
}

export interface VoterLogin {
  /** Voter ID or email */
  identifier: string;
  password: string;
  rememberMe?: boolean;
}

export interface OtpVerification {
  voterId: string;
  otp: string;
}

export interface ResendOtp {
  voterId: string;
}

export interface AdminLogin {
  username: string;
  password: string;
}

export type VoterRole = typeof VoterRole[keyof typeof VoterRole];


export const VoterRole = {
  voter: 'voter',
  admin: 'admin',
} as const;

export interface Voter {
  id: number;
  fullName: string;
  voterId: string;
  email: string;
  mobile: string;
  hasVoted: boolean;
  role: VoterRole;
}

export interface AuthResponse {
  token: string;
  user: Voter;
}

export interface LoginResponse {
  message: string;
  voterId: string;
}

export interface VoterSummary {
  id: number;
  fullName: string;
  voterId: string;
  email: string;
  mobile: string;
  hasVoted: boolean;
  createdAt: string;
}

export type ElectionStatus = typeof ElectionStatus[keyof typeof ElectionStatus];


export const ElectionStatus = {
  upcoming: 'upcoming',
  active: 'active',
  ended: 'ended',
} as const;

export interface Election {
  id: number;
  title: string;
  description: string;
  startDate: string;
  endDate: string;
  status: ElectionStatus;
  createdAt?: string;
}

export interface ElectionInput {
  title: string;
  description: string;
  startDate: string;
  endDate: string;
  status?: string;
}

export interface ElectionUpdate {
  title?: string;
  description?: string;
  startDate?: string;
  endDate?: string;
  status?: string;
}

export interface Candidate {
  id: number;
  name: string;
  party: string;
  symbol: string;
  description: string;
  imageUrl: string;
  electionId: number;
}

export interface CandidateInput {
  name: string;
  party: string;
  symbol: string;
  description: string;
  imageUrl?: string;
  electionId: number;
}

export interface CandidateUpdate {
  name?: string;
  party?: string;
  symbol?: string;
  description?: string;
  imageUrl?: string;
}

export interface VoteInput {
  candidateId: number;
  electionId: number;
}

export interface VoteRecord {
  id: number;
  voterId: number;
  candidateId: number;
  electionId: number;
  timestamp: string;
}

export interface VoteStatus {
  hasVoted: boolean;
  /** @nullable */
  candidateId?: number | null;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  timestamp?: string | null;
}

export interface CandidateResult {
  id: number;
  name: string;
  party: string;
  symbol: string;
  imageUrl: string;
  voteCount: number;
  percentage: number;
}

export interface ElectionResults {
  election: Election;
  candidates: CandidateResult[];
  totalVotes: number;
  winner: CandidateResult;
}

export interface AdminStats {
  totalVoters: number;
  totalVoted: number;
  totalCandidates: number;
  totalElections: number;
  votingPercentage: number;
  /** @nullable */
  activeElection: string | null;
}

export interface ActivityLog {
  id: number;
  action: string;
  description: string;
  timestamp: string;
  /** @nullable */
  userId: number | null;
  userType?: string;
}