# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SecureVote API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerVoter
      tags: [auth]
      summary: Register a new voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterRegistration"
      responses:
        "201":
          description: Voter registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Voter ID or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginVoter
      tags: [auth]
      summary: Login as voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterLogin"
      responses:
        "200":
          description: Login successful - OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerification"
      responses:
        "200":
          description: OTP verified, token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      summary: Resend OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendOtp"
      responses:
        "200":
          description: OTP resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current voter profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Voter profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voter"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: loginAdmin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLogin"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates for active election
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a specific candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Candidate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/voters:
    get:
      operationId: listVoters
      tags: [admin]
      summary: List all registered voters (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of voters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VoterSummary"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/activity:
    get:
      operationId: getActivityLog
      tags: [admin]
      summary: Get activity logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activity log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    VoterRegistration:
      type: object
      required: [fullName, voterId, email, mobile, password]
      properties:
        fullName:
          type: string
          minLength: 2
        voterId:
          type: string
          minLength: 4
        email:
          type: string
          format: email
        mobile:
          type: string
        password:
          type: string
          minLength: 8
    VoterLogin:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Voter ID or email
        password:
          type: string
        rememberMe:
          type: boolean
    OtpVerification:
      type: object
      required: [voterId, otp]
      properties:
        voterId:
          type: string
        otp:
          type: string
    ResendOtp:
      type: object
      required: [voterId]
      properties:
        voterId:
          type: string
    AdminLogin:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/Voter"
    LoginResponse:
      type: object
      required: [message, voterId]
      properties:
        message:
          type: string
        voterId:
          type: string
    Voter:
      type: object
      required: [id, fullName, voterId, email, mobile, hasVoted, role]
      properties:
        id:
          type: integer
        fullName:
          type: string
        voterId:
          type: string
        email:
          type: string
        mobile:
          type: string
        hasVoted:
          type: boolean
        role:
          type: string
          enum: [voter, admin]
    VoterSummary:
      type: object
      required: [id, fullName, voterId, email, mobile, hasVoted, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        voterId:
          type: string
        email:
          type: string
        mobile:
          type: string
        hasVoted:
          type: boolean
        createdAt:
          type: string
...
    Candidate:
      type: object
      required: [id, name, party, symbol, description, imageUrl, electionId]
      properties:
        id:
          type: integer
        name:
          type: string
        party:
          type: string
        symbol:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        electionId:
          type: integer
    CandidateInput:
      type: object
      required: [name, party, symbol, description, electionId]
      properties:
        name:
          type: string
        party:
          type: string
        symbol:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        electionId:
          type: integer
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        symbol:
          type: string
        description:
          type: string
        imageUrl:
          type: string
...
    AdminStats:
      type: object
      required: [totalVoters, totalVoted, totalCandidates, totalElections, votingPercentage, activeElection]
      properties:
        totalVoters:
          type: integer
        totalVoted:
          type: integer
        totalCandidates:
          type: integer
        totalElections:
          type: integer
        votingPercentage:
          type: number
        activeElection:
          type: ["string", "null"]
    ActivityLog:
      type: object
      required: [id, action, description, timestamp, userId]
      properties:
        id:
          type: integer
        action:
          type: string
        description:
          type: string
        timestamp:
          type: string
        userId:
          type: ["integer", "null"]
        userType:
          type: string