# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  action: string;
  description: string;
  timestamp: string;
  /** @nullable */
  userId: number | null;
  userType?: string;
}