# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalVoters: number;
  totalVoted: number;
  totalCandidates: number;
  totalElections: number;
  votingPercentage: number;
  /** @nullable */
  activeElection: string | null;
}