# lib/api-zod/src/generated/types/candidateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateInput {
  name: string;
  party: string;
  symbol: string;
  description: string;
  imageUrl?: string;
  electionId: number;
}