# lib/api-zod/src/generated/types/candidateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateUpdate {
  name?: string;
  party?: string;
  symbol?: string;
  description?: string;
  imageUrl?: string;
}