# lib/api-zod/src/generated/types/voterLogin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SecureVote API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoterLogin {
  /** Voter ID or email */
  identifier: string;
  password: string;
  rememberMe?: boolean;
}